ner_training/requirements.txt
# requirements.txt
torch
transformers
datasets
seqeval
pydantic
python-dateutil
fastapi
uvicorn[standard]
mlflow